restructure routes
