+++ src/routing/messages/messages.route.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useMemo } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
-import { FourOuFour, Messages as MessagesPage } from '../pages';
+import { FourOuFour, Messages as MessagesPage } from '../../pages';
 
 export const Messages: FC = () => {
   const { path, url } = useRouteMatch();
