+++ src/routing/routing.tsx
@@ -7,12 +7,12 @@ import HomePage from '../pages/home/Home';
 import { AuthorizationCredential } from '../models/graphql-schema';
 import { AdminRoute } from './admin/admin.route';
 import { IdentityRoute } from './identity/identity';
-import { EcoverseRoute } from './ecoverse.route';
-import { Messages } from './messages';
+import { EcoverseRoute } from './ecoverse/ecoverse.route';
+import { Messages } from './messages/messages.route';
 import { Restricted } from './restricted';
 import RestrictedRoute from './route.extensions';
 import { SearchRoute } from './search.route';
-import OrganizationRoute from './organization.route';
+import OrganizationRoute from './organization/organization.route';
 import { UserRoute } from './user/user';
 import ProfilePage from '../pages/ProfilePage';
 import { OrganizationProvider } from '../context/OrganizationProvider';
